Move format_bytes to chia.cmds.cmds_util
