+++ chia/cmds/beta_funcs.py
@@ -5,10 +5,11 @@ import sys
 from pathlib import Path
 from typing import Any, Callable, Dict, List, Optional
 
+from chia.cmds.cmds_util import format_bytes
 from chia.util.beta_metrics import metrics_log_interval_max, metrics_log_interval_min
 from chia.util.chia_logging import get_beta_logging_config
 from chia.util.errors import InvalidPathError
-from chia.util.misc import format_bytes, prompt_yes_no, validate_directory_writable
+from chia.util.misc import prompt_yes_no, validate_directory_writable
 
 
 def default_beta_root_path() -> Path:
