+++ chia/cmds/cmds_util.py
@@ -362,3 +362,18 @@ class CMDTXConfigLoader(CMDCoinSelectionConfigLoader):
             cs_config.excluded_coin_ids,
             self.reuse_puzhash,
         ).autofill(constants=DEFAULT_CONSTANTS, config=config, logged_in_fingerprint=fingerprint)
+
+
+def format_bytes(bytes: int) -> str:
+    if not isinstance(bytes, int) or bytes < 0:
+        return "Invalid"
+
+    LABELS = ("MiB", "GiB", "TiB", "PiB", "EiB", "ZiB", "YiB")
+    BASE = 1024
+    value = bytes / BASE
+    for label in LABELS:
+        value /= BASE
+        if value < BASE:
+            return f"{value:.3f} {label}"
+
+    return f"{value:.3f} {LABELS[-1]}"
