+++ chia/cmds/show_funcs.py
@@ -11,9 +11,9 @@ from chia.types.blockchain_format.sized_bytes import bytes32
 async def print_blockchain_state(node_client: FullNodeRpcClient, config: Dict[str, Any]) -> bool:
     import time
 
+    from chia.cmds.cmds_util import format_bytes
     from chia.consensus.block_record import BlockRecord
     from chia.util.ints import uint64
-    from chia.util.misc import format_bytes
 
     blockchain_state = await node_client.get_blockchain_state()
     if blockchain_state is None:
